+++ examples/cloneless.rs
@@ -10,8 +10,8 @@ use std::{
 use rand::Rng;
 use serde::{Deserialize, Serialize};
 use warqueen::{
-    ClientEvent, ClientNetworking, ClientOnServerEvent, ClonelessSending, DisconnectionDetails,
-    NetReceive, NetSend, ServerListenerNetworking,
+    ClientDisconnectionDetails, ClientEvent, ClientNetworking, ClientOnServerDisconnectionDetails,
+    ClientOnServerEvent, ClonelessSending, NetReceive, NetSend, ServerListenerNetworking,
 };
 
 // See `example_client.rs` and `example_server.rs` for more comments.
@@ -98,8 +98,15 @@ fn client() {
                         println!("The server says \"{content}\"")
                     }
                 },
-                ClientEvent::Disconnected => {
-                    println!("Server disconnected, let's terminate");
+                ClientEvent::Disconnected(details) => {
+                    match details {
+                        ClientDisconnectionDetails::None => {
+                            println!("Server disconnected, let's terminate");
+                        }
+                        ClientDisconnectionDetails::Timeout => {
+                            println!("Server timed out, let's terminate");
+                        }
+                    }
                     return;
                 }
                 ClientEvent::FailedToConnect => {
@@ -183,10 +190,10 @@ fn server() {
                     },
                     ClientOnServerEvent::Disconnected(details) => {
                         match details {
-                            DisconnectionDetails::None => {
+                            ClientOnServerDisconnectionDetails::None => {
                                 println!("Client {client_id} disconnected")
                             }
-                            DisconnectionDetails::Timeout => {
+                            ClientOnServerDisconnectionDetails::Timeout => {
                                 println!("Client {client_id} timed out")
                             }
                         }